Fix LocalMusicRepository names
